Moved all frames/pages into one python file
All gui related stuff in one python file. One small fix to userAccountPage for path finding the avatar image due to change. openPage function changed, it relies on a map now rather than importing from different python modules
