+++ Main.py
@@ -1,26 +1,498 @@
-# For loading different pages 
-import importlib
-
 # Tkinter itself
 import tkinter as tk
 import customtkinter as ctk 
-
 # Import our models
 from classes.models import *
-
+# import utility functions
+from classes.utilities import *
+# Import hash library for hashing passwords
+import hashlib
 # Import sqlalchemy to do our operations
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
-
-
 # For creating a dynamic footer 
 import datetime
+# For image rendering
+from PIL import Image, ImageTk
+import os			
 
 
 
 ctk.set_appearance_mode("Light")  # Modes: "System" (standard), "Dark", "Light"
 
-# Header class that has a navbar
+# class homePage(ctk.CTkFrame):
+# 	def __init__(self, master):
+# 		super().__init__(master)
+# 		self.master = master
+
+# class aiSettingsPage(ctk.CTkFrame):
+# 	def __init__(self, master):
+# 		super().__init__(master)
+# 		self.master = master
+
+class storyLibraryPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+
+		
+
+
+
+
+##### Page for changing passwords #####
+class changePasswordPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+
+		# Create edit form frame
+		form = ctk.CTkFrame(self)
+		form.pack(expand=True)
+
+		# Create header of the form
+		formHeader = ctk.CTkFrame(form, fg_color="transparent")
+		formHeading = ctk.CTkLabel(formHeader, text="Change Password", font=("Helvetica", 32))
+		
+		passwordRulesMessage = ctk.CTkLabel(formHeader, text="Password has to be some long and have numbers!")
+		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
+		
+		# Create the input section with form fields 
+		formFieldsSection = ctk.CTkFrame(form)
+
+		formFields = [
+			"Old Password",
+			"New Password",
+			"Retype New Password",
+		]
+		# Create list of form entries to get input later
+		# Then create and position label and entry widgets for form
+		self.formEntryList = []
+		for x in range(len(formFields)):
+			label = ctk.CTkLabel(formFieldsSection, text=formFields[x])
+			entry = ctk.CTkEntry(formFieldsSection)
+			checkVar = ctk.StringVar(value="off")
+			visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
+			self.formEntryList.append(entry)
+			label.grid(row=x, column=0, pady=10, padx=10)
+			entry.grid(row=x, column=1, pady=10, padx=10)
+			visibilityCheckBox.grid(row=x, column=2, pady=10, padx=4)
+
+		# Create section to have form buttons/actions
+		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
+		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))		
+		changePasswordBtn = ctk.CTkButton(formBtnsSection, text="Confirm Change")
+
+		# Structure the remaining elements of the page
+		# padx=90; adds enough gray padding so that form looks uniform
+		formHeader.grid(row=0, column=0, pady=10, padx=90)
+		formHeading.grid(row=0, column=0)
+		passwordRulesMessage.grid(row=1, column=0)
+		self.formErrorMessage.grid(row=2, column=0)
+		formFieldsSection.grid(row=1, column=0, pady=10)
+		formBtnsSection.grid(row=2, column=0, pady=10)
+		clearFormBtn.grid(row=0, column=0, padx=10, pady=10)
+		changePasswordBtn.grid(row=0, column=1, padx=10, pady=10)
+
+
+
+##### Delete Account Page #####
+class deleteAccountPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+		# Create edit form frame
+		form = ctk.CTkFrame(self)
+		form.pack(expand=True)
+		# Create header of the form
+		formHeader = ctk.CTkFrame(form, fg_color="transparent")
+		formHeading = ctk.CTkLabel(formHeader, text="Delete Account", font=("Helvetica", 32))
+		subHeading = ctk.CTkLabel(formHeader, text="Are you sure you want to delete 'Insert Username'?")
+		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
+		
+		# Create the input section with form fields 
+		formFieldsSection = ctk.CTkFrame(form)
+
+		formFields = [
+			{
+				"text": "Username"
+			},
+			{
+				"text": "Password", 
+				"toggleHidden": True
+			}
+		]
+		# Create list of form entries to get input later
+		# Then create and position label and entry widgets for form
+		self.formEntryList = []
+
+		for x in range(len(formFields)):
+			label = ctk.CTkLabel(formFieldsSection, text=formFields[x].get("text"))
+			entry = ctk.CTkEntry(formFieldsSection)
+			self.formEntryList.append(entry)
+			label.grid(row=x, column=0, padx=10, pady=10)
+			entry.grid(row=x, column=1, padx=10, pady=10)
+			if (formFields[x].get("toggleHidden")):
+				checkVar = ctk.StringVar(value="off")
+				visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
+				visibilityCheckBox.grid(row=x, column=2, padx=4, pady=10)
+
+		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
+		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
+		deleteAccountBtn = ctk.CTkButton(formBtnsSection, text="Delete Account")
+		
+		# Structure the remaining elements of the page
+		formHeader.grid(row=0, column=0, padx=40, pady=10)
+		formHeading.grid(row=0, column=0)
+		subHeading.grid(row=1, column=0)
+		self.formErrorMessage.grid(row=2, column=0)
+		formFieldsSection.grid(row=1, column=0, pady=10)
+		formBtnsSection.grid(row=2, column=0, pady=10)
+		clearFormBtn.grid(row=0,column=0, padx=10, pady=10)
+		deleteAccountBtn.grid(row=0,column=1, padx=10, pady=10)
+
+
+
+###### The page for editting accounts #####
+class editAccountPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+
+		# Create edit form frame
+		form = ctk.CTkFrame(self)
+		form.pack(expand=True)
+
+		# Create header of the form
+		formHeader = ctk.CTkFrame(form, fg_color="transparent")
+		formHeading = ctk.CTkLabel(formHeader, text="Edit Account", font=("Helvetica", 32))
+		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
+		
+		# Create the input section with form fields 
+		formFieldsSection = ctk.CTkFrame(form)
+
+		formFields = [
+			"Username",
+			"Email",
+			"First Name",
+			"Last Name"
+		]
+		# Create list of form entries to get input later
+		# Then create and position label and entry widgets for form
+		self.formEntryList = []
+		for x in range(len(formFields)):
+			label = ctk.CTkLabel(formFieldsSection, text=formFields[x])
+			entry = ctk.CTkEntry(formFieldsSection)
+			label.grid(row=x, column=0, pady=10, ipadx=10)
+			entry.grid(row=x, column=1, pady=10, ipadx=10)
+			self.formEntryList.append(entry)
+
+		# Clear form button and then a confirm changes button
+		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
+		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
+		confirmEditsBtn = ctk.CTkButton(formBtnsSection, text="Confirm Edits")
+		openChangePasswordBtn = ctk.CTkButton(formBtnsSection, text="Change Password", command=lambda: self.master.openPage("changePasswordPage"))
+		openDeleteAccountBtn = ctk.CTkButton(formBtnsSection, text="Account Deletion", command=lambda: self.master.openPage("deleteAccountPage")) #type: ignore
+
+		# Structure the remaining page elements accordingly
+		formHeader.grid(row=0, column=0, padx=70, pady=10)
+		formHeading.grid(row=0, column=0)
+		self.formErrorMessage.grid(row=1, column=0)
+		formFieldsSection.grid(row=1, column=0, pady=10, ipadx=10)
+		formBtnsSection.grid(row=2, column=0, pady=10)
+		confirmEditsBtn.grid(row=0, column=0, padx=10, pady=10)
+		clearFormBtn.grid(row=1, column=0, padx=10, pady=10)
+		openChangePasswordBtn.grid(row=2, column=0, padx=10, pady=10)
+		openDeleteAccountBtn.grid(row=3, column=0, padx=10, pady=10)
+
+
+
+##### The user account or profile page#####
+class userAccountPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+		# This extra frame just allows stuff to be centered onto the userAccountPage frame
+		innerPageFrame = ctk.CTkFrame(self, fg_color="transparent")
+		innerPageFrame.pack(expand=True)
+		# Create section to store the user's profile picture
+		userImageSection = tk.Canvas(innerPageFrame, highlightbackground="#eee", highlightthickness=1)
+		
+		avatarSourcePath = f"./assets/images/{self.master.loggedInUser.avatar}"
+		image = Image.open(avatarSourcePath).resize((500, 500))
+		imageWidget = ImageTk.PhotoImage(image=image)
+		imageLabel = tk.Label(userImageSection, image=imageWidget)
+		imageLabel.image = imageWidget #type: ignore
+		imageLabel.grid(row=0, column=0)
+
+		# Create section to store buttons on the user page
+		userBtnsSection = ctk.CTkFrame(innerPageFrame, fg_color="transparent")
+		openEditAccountBtn = ctk.CTkButton(userBtnsSection, text="Edit Account", command=lambda: self.master.openPage("editAccountPage")) #type: ignore
+		confirmLogOutBtn = ctk.CTkButton(userBtnsSection, text="Log Out", command=self.logoutUser)
+		openEditAccountBtn.grid(row=0, column=0, pady=5)
+		confirmLogOutBtn.grid(row=1, column=0, pady=5)
+
+		# Create section to display user information
+		# Get user information, and iteratively create labels to show that user information
+		userInfoSection = ctk.CTkFrame(innerPageFrame, fg_color="transparent")
+		userInfoFields = [
+			{
+				"text": "Username",
+				"value": self.master.loggedInUser.username
+			},
+			{
+				"text": "Email",
+				"value": self.master.loggedInUser.email
+			},
+			{
+				"text": "Name",
+				"value": f"{self.master.loggedInUser.firstName} {self.master.loggedInUser.lastName}"
+			}
+		]
+		for x in range(len(userInfoFields)):
+			label = ctk.CTkLabel(userInfoSection, text=f"{userInfoFields[x].get('text')}: {userInfoFields[x].get('value')}", font=("Helvetica", 24))
+			label.grid(row=x, column=0, sticky="W", pady=10)
+
+		# Structure the 3 main sections of the user account page
+		userImageSection.grid(row=0, column=0, padx=30, pady=10)
+		userBtnsSection.grid(row=1, padx=30, column=0)
+		userInfoSection.grid(row=0, column=1, padx=30)
+
+	# Log the current user out of the application
+	def logoutUser(self):
+		# loggedInUser is none because the user is logging out 
+		self.master.loggedInUser = None #type: ignore
+		# Redirect the user to the login page
+		self.master.openPage("userLoginPage") #type: ignore
+		# Update nav buttons so that user can't access the pages associated with them
+		self.master.header.updateNavButtons() #type: ignore
+
+
+
+##### User Registration Page ######
+class userRegisterPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+		
+		# Create registration form frame
+		form = ctk.CTkFrame(self)
+		form.pack(expand=True)
+
+		# Create form header and text elements for header
+		formHeader = ctk.CTkFrame(form, fg_color="transparent")
+		formHeading = ctk.CTkLabel(formHeader, text="Register", font=("Helvetica", 32))
+		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
+
+		# Create input section with form fields
+		formFieldsSection = ctk.CTkFrame(form)
+		formFields = [
+			{
+				"text": "Email",
+			},
+			{
+				"text": "Username",
+			},
+			{
+				"text": "First Name",
+			},
+			{
+				"text": "Last Name",
+			},
+			{
+				"text": "Password",
+				"toggleHidden": True	
+			},
+			{
+				"text": "Confirm Password",
+				"toggleHidden": True
+			},
+		]		
+		# Create list of form entries to get input later
+		self.formEntryList = []
+		# Iterate through object to create fields
+		for x in range(len(formFields)):
+			label = ctk.CTkLabel(formFieldsSection, text=formFields[x]["text"])
+			entry = ctk.CTkEntry(formFieldsSection)
+			self.formEntryList.append(entry)
+			label.grid(row=x, column=0, padx=10, pady=10)
+			entry.grid(row=x, column=1, padx=10, pady=10)
+			# If there's a 'toggleHidden' attribute, add a checkbox
+			# so that we can toggle visibility on the field
+			if (formFields[x].get("toggleHidden")):
+				checkVar = ctk.StringVar(value="off")
+				visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
+				visibilityCheckBox.grid(row=x, column=2, padx=4, pady=10)
+			
+		# Create section to have form buttons/actions
+		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
+		openLoginBtn = ctk.CTkButton(formBtnsSection, text="Log into an existing account", command=lambda: self.master.openPage("userLoginPage")) #type: ignore
+		confirmRegisterBtn = ctk.CTkButton(formBtnsSection, text="Confirm Registration", command=self.registerUser)
+		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
+
+		# Structure the remaining elements of the page
+		formHeader.grid(row=0, column=0, pady=10)
+		formHeading.grid(row=0, column=0)
+		self.formErrorMessage.grid(row=1, column=0)
+		formFieldsSection.grid(row=1, column=0, pady=10, ipadx=10)
+		formBtnsSection.grid(row=2, column=0, pady=10)
+		clearFormBtn.grid(row=0, column=0, padx=10, pady=10)
+		openLoginBtn.grid(row=0, column=1, padx=10, pady=10)
+		confirmRegisterBtn.grid(row=0, column=2, padx=10, pady=10)
+
+	# Registers a user in the database
+	def registerUser(self):
+		# Check if any fields are empty before moving on
+		if isEmptyEntryWidgets(self.formEntryList):
+			self.formErrorMessage.configure(text="Some fields are empty!")
+			return
+
+		# Get input field values
+		email = self.formEntryList[0].get()
+		username = self.formEntryList[1].get()
+		firstName = self.formEntryList[2].get()
+		lastName = self.formEntryList[3].get()
+		password = self.formEntryList[4].get()
+		confirmPassword = self.formEntryList[5].get()
+
+		# Check if it's a valid email form
+		if not isValidEmail(email):
+			self.formErrorMessage.configure(text="Email isn't in valid form!")
+			return
+		
+		# Check if username is valid
+		if not isValidUsername(username):
+			self.formErrorMessage.configure(text="Username length is 6-20 characters, and can only have numbers, letters, and underscores!")
+			return
+		
+		# Check if password is valid
+		if not isValidPassword(password):
+			self.formErrorMessage.configure(text="Password can only be 6-20 characters with only numbers, letters, and symbols: !@#$%^&*(){}[]<>,+~-._")
+			return
+		
+		# Check if password and the retryped password match
+		if password != confirmPassword:
+			self.formErrorMessage.configure(text="Passwords must match!")
+			return
+		
+		# Create a session that we can use
+		session = self.master.Session() #type: ignore
+
+		# Check if there are any users with the inputted username
+		retrievedUser = session.query(User).filter_by(username=username).first()
+		if retrievedUser:
+			self.formErrorMessage.configure(text="Usename already taken!")
+			return
+		
+		# Create a new user based on the form information
+		newUser = User(
+			email=email,
+			username=username,
+			firstName=firstName,
+			lastName=lastName,
+			# For right now do a simple md5 hash, but later
+			# we should use that api idea
+			passwordHash=hashlib.md5(password.encode("utf-8")).hexdigest(),
+			avatar="default_user.jpg"
+		)
+
+		# Add new user to the database
+		session.add(newUser)
+		session.commit()
+		session.close()
+		# Redirect user to login screen after they've successfully registered
+		self.master.openPage("userLoginPage") #type: ignore
+
+
+
+##### User login page #####
+class userLoginPage(ctk.CTkFrame):
+	def __init__(self, master):
+		super().__init__(master)
+		self.master = master
+		# Create login form frame
+		form = ctk.CTkFrame(self)
+		form.pack(expand=True)
+
+		# Create header of the form
+		formHeader = ctk.CTkFrame(form, fg_color="transparent")
+		formHeading = ctk.CTkLabel(formHeader, text="Login", font=("Helvetica", 32))
+		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
+		
+		# Create the input section with form fields 
+		formFieldsSection = ctk.CTkFrame(form)
+
+		formFields = [
+			{
+				"text": "Username"
+			},
+			{
+				"text": "Password",
+				"toggleHidden": True,
+			}
+		]
+		# Create list of form entries to get input later
+		self.formEntryList = []
+		for x in range(len(formFields)):
+			label = ctk.CTkLabel(formFieldsSection, text=formFields[x].get("text"))
+			entry = ctk.CTkEntry(formFieldsSection)
+			label.grid(row=x, column=0, pady=10, ipadx=10)
+			entry.grid(row=x, column=1, pady=10, ipadx=10)
+			if (formFields[x].get("toggleHidden")):
+				checkVar = ctk.StringVar(value="off")
+				visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
+				visibilityCheckBox.grid(row=x, column=2, padx=4, pady=10)
+			self.formEntryList.append(entry)
+
+		# Create section to have form buttons/actions
+		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
+		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
+		confirmLoginBtn = ctk.CTkButton(formBtnsSection, text="Confirm Login", command=self.loginUser)
+		openRegisterAccountBtn = ctk.CTkButton(formBtnsSection, text="Register New Account", command=lambda: self.master.openPage("userRegisterPage")) #type: ignore
+		
+		# Structure the remaining elements of the page
+		formHeader.grid(row=0, column=0, pady=10)
+		formHeading.grid(row=0, column=0)
+		self.formErrorMessage.grid(row=1, column=0)
+		formFieldsSection.grid(row=1, column=0, pady=10, ipadx=10)
+		formBtnsSection.grid(row=2, column=0, pady=10)
+		clearFormBtn.grid(row=0, column=0, padx=10, pady=10)
+		openRegisterAccountBtn.grid(row=0, column=1, padx=10, pady=10)
+		confirmLoginBtn.grid(row=0, column=2, padx=10, pady=10)
+		
+
+	## Attempts to log in a user
+	def loginUser(self):
+		# Check if input fields are empty
+		if (isEmptyEntryWidgets(self.formEntryList)):
+			self.formErrorMessage.configure(text="Some fields are empty!")
+			return
+		
+		# Get input values from the form
+		username = self.formEntryList[0].get()
+		password = self.formEntryList[1].get()
+		passwordHash = hashlib.md5(password.encode("utf-8")).hexdigest()
+
+		# Now check if the inputted username and password hash matches a record from the User table 
+		session = self.master.Session() #type: ignore
+		retrievedUser = session.query(User).filter_by(username=username, passwordHash=passwordHash).first()
+		if not retrievedUser:
+			self.formErrorMessage.configure(text="Username or password is incorrect!")
+			return
+		
+		# Assign the new logged in user
+		self.master.loggedInUser = retrievedUser #type: ignore
+
+		# Update the nav buttons now that the user is logged in
+		# so that they actually work and aren't disabled
+		self.master.header.updateNavButtons() #type: ignore
+
+		# Redirect the user to the 'My Account' or the 'user account page'
+		self.master.openPage("userAccountPage") #type: ignore
+
+
+
+##### Application Header that has navbar #####
 class Header(ctk.CTkFrame):
 	def __init__(self, master):
 		super().__init__(master, fg_color="#f9a8d4", corner_radius=0)
@@ -55,7 +527,6 @@ class Header(ctk.CTkFrame):
 
 		# Adjust nav button links depending on user login state
 		self.updateNavButtons()
-
 	'''
 	- Update the nav buttons based on the login state of the user
 	1. If user is logged in, nav buttons lead to their respective places
@@ -67,8 +538,10 @@ class Header(ctk.CTkFrame):
 				button.configure(state="standard")
 			else:
 				button.configure(state="disabled")
-			
-# Footer frame
+
+
+
+###### Application Footer #####
 class Footer(ctk.CTkFrame):
 	def __init__(self, master):
 		super().__init__(master, fg_color="#f9a8d4", corner_radius=0)
@@ -76,7 +549,9 @@ class Footer(ctk.CTkFrame):
 		footerLabel = ctk.CTkLabel(self, text=f"BookSmart {currentYear}", text_color="black",	font=("Helvetica", 16))
 		footerLabel.pack()
 
-# Main application
+
+
+##### Main application #####
 class App(ctk.CTk):
 	def __init__(self):
 		super().__init__()
@@ -92,10 +567,51 @@ class App(ctk.CTk):
 		# the User that is currently logged into the application
 		self.loggedInUser = None 
 
+		# Frames in the application
+		self.pageMap = {
+			"userLoginPage": userLoginPage,
+			"userRegisterPage": userRegisterPage,
+			"userAccountPage": userAccountPage,
+			"editAccountPage": editAccountPage,
+			"deleteAccountPage": deleteAccountPage,
+			"changePasswordPage": changePasswordPage,
+		}
+
 		# Engine and session constructor that we're going to use 
 		self.engine = create_engine("sqlite:///assets/PyProject.db")
 		self.Session = sessionmaker(bind=self.engine)
-		
+
+
+
+
+
+
+
+
+
+
+		session = self.Session()
+		self.loggedInUser = session.query(User).filter_by(username="knguyen44").first()
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 		# Call function to create navbar
 		self.header = Header(self)
 		self.header.pack(side="top", fill="x")
@@ -103,27 +619,19 @@ class App(ctk.CTk):
 		footer = Footer(self)
 		footer.pack(fill="x", side="bottom")
 
-		# Open the account page for test purposes
-		self.openPage("userLoginPage")
-
-
-	def loadPage(self, pageName):
-		try:
-			module = importlib.import_module(f"pages.{pageName}")
-			pageClass = getattr(module, pageName)
-			return pageClass		
-		except (ImportError, AttributeError):
-			print(f"Error: Page {pageName} wasn't found.")
-			return None
+		# Start off by making the user log in
+		self.openPage("userAccountPage")
 		
 	def openPage(self, pageName, *args):
-		pageClass = self.loadPage(pageName)
-		if pageClass is None:
-			return
-		if self.currentPage:
-			self.currentPage.destroy()
-		self.currentPage = pageClass(self, *args)
-		self.currentPage.pack(fill="both", expand=True)
+		try:
+			pageClass = self.pageMap[pageName]
+			if self.currentPage:
+				self.currentPage.destroy()
+			self.currentPage = pageClass(self, *args)
+			self.currentPage.pack(fill="both", expand=True)
+		except KeyError:
+			print(f"Error: Page {pageName} doesn't exist")
+
 
 if __name__ == "__main__":
 	app = App()
