--- pages/changePasswordPage.py
@@ -1,61 +1,0 @@
-import customtkinter as ctk
-import sys
-sys.path.append("..")
-from classes.utilities import *
-
-class changePasswordPage(ctk.CTkFrame):
-	def __init__(self, master):
-		super().__init__(master)
-		self.master = master
-
-		# Create edit form frame
-		form = ctk.CTkFrame(self)
-		form.pack(expand=True)
-
-		# Create header of the form
-		formHeader = ctk.CTkFrame(form, fg_color="transparent")
-		formHeading = ctk.CTkLabel(formHeader, text="Change Password", font=("Helvetica", 32))
-		
-		passwordRulesMessage = ctk.CTkLabel(formHeader, text="Password has to be some long and have numbers!")
-		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
-		
-		# Create the input section with form fields 
-		formFieldsSection = ctk.CTkFrame(form)
-
-		formFields = [
-			"Old Password",
-			"New Password",
-			"Retype New Password",
-		]
-		# Create list of form entries to get input later
-		# Then create and position label and entry widgets for form
-		self.formEntryList = []
-		
-		for x in range(len(formFields)):
-			label = ctk.CTkLabel(formFieldsSection, text=formFields[x])
-			entry = ctk.CTkEntry(formFieldsSection)
-			checkVar = ctk.StringVar(value="off")
-			visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
-			self.formEntryList.append(entry)
-			label.grid(row=x, column=0, pady=10, padx=10)
-			entry.grid(row=x, column=1, pady=10, padx=10)
-			visibilityCheckBox.grid(row=x, column=2, pady=10, padx=4)
-
-
-		# Create section to have form buttons/actions
-		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
-		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))		
-		changePasswordBtn = ctk.CTkButton(formBtnsSection, text="Confirm Change")
-
-		# Structure the remaining elements of the page
-		# padx=90; adds enough gray padding so that form looks uniform
-		formHeader.grid(row=0, column=0, pady=10, padx=90)
-		formHeading.grid(row=0, column=0)
-		passwordRulesMessage.grid(row=1, column=0)
-		self.formErrorMessage.grid(row=2, column=0)
-
-		formFieldsSection.grid(row=1, column=0, pady=10)
-		
-		formBtnsSection.grid(row=2, column=0, pady=10)
-		clearFormBtn.grid(row=0, column=0, padx=10, pady=10)
-		changePasswordBtn.grid(row=0, column=1, padx=10, pady=10)
