--- pages/deleteAccountPage.py
@@ -1,70 +1,0 @@
-import customtkinter as ctk
-import sys
-sys.path.append("..")
-from classes.utilities import *
-
-class deleteAccountPage(ctk.CTkFrame):
-	def __init__(self, master):
-		super().__init__(master)
-		self.master = master
-
-		# Create edit form frame
-		form = ctk.CTkFrame(self)
-		form.pack(expand=True)
-
-		# Create header of the form
-		formHeader = ctk.CTkFrame(form, fg_color="transparent")
-		formHeading = ctk.CTkLabel(formHeader, text="Delete Account", font=("Helvetica", 32))
-		subHeading = ctk.CTkLabel(formHeader, text="Are you sure you want to delete 'Insert Username'?")
-		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
-		
-		# Create the input section with form fields 
-		formFieldsSection = ctk.CTkFrame(form)
-
-		formFields = [
-			{
-				"text": "Username"
-			},
-			{
-				"text": "Password", 
-				"toggleHidden": True
-			}
-		]
-		# Create list of form entries to get input later
-		# Then create and position label and entry widgets for form
-		self.formEntryList = []
-
-		for x in range(len(formFields)):
-			label = ctk.CTkLabel(formFieldsSection, text=formFields[x].get("text"))
-			entry = ctk.CTkEntry(formFieldsSection)
-			self.formEntryList.append(entry)
-			label.grid(row=x, column=0, padx=10, pady=10)
-			entry.grid(row=x, column=1, padx=10, pady=10)
-			if (formFields[x].get("toggleHidden")):
-				checkVar = ctk.StringVar(value="off")
-				visibilityCheckBox = ctk.CTkCheckBox(formFieldsSection, variable=checkVar, command=lambda entry=entry, var=checkVar: toggleHidden(entry, var),  text="Hide", onvalue="on", offvalue="off")
-				visibilityCheckBox.grid(row=x, column=2, padx=4, pady=10)
-
-			formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
-			clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
-			deleteAccountBtn = ctk.CTkButton(formBtnsSection, text="Delete Account")
-			
-
-		# Structure the remaining elements of the page
-		formHeader.grid(row=0, column=0, padx=40, pady=10)
-		formHeading.grid(row=0, column=0)
-		subHeading.grid(row=1, column=0)
-		self.formErrorMessage.grid(row=2, column=0)
-
-		formFieldsSection.grid(row=1, column=0, pady=10)
-
-		formBtnsSection.grid(row=2, column=0, pady=10)
-		clearFormBtn.grid(row=0,column=0, padx=10, pady=10)
-		deleteAccountBtn.grid(row=0,column=1, padx=10, pady=10)
-
-		
-			
-			
-			
-
-
