--- pages/editAccountPage.py
@@ -1,58 +1,0 @@
-import tkinter 
-import customtkinter as ctk
-import sys
-sys.path.append("..")
-from classes.utilities import *
-
-
-class editAccountPage(ctk.CTkFrame):
-	def __init__(self, master):
-		super().__init__(master)
-		self.master = master
-
-		# Create edit form frame
-		form = ctk.CTkFrame(self)
-		form.pack(expand=True)
-
-		# Create header of the form
-		formHeader = ctk.CTkFrame(form, fg_color="transparent")
-		formHeading = ctk.CTkLabel(formHeader, text="Edit Account", font=("Helvetica", 32))
-		self.formErrorMessage = ctk.CTkLabel(formHeader, text="")
-		
-		# Create the input section with form fields 
-		formFieldsSection = ctk.CTkFrame(form)
-
-		formFields = [
-			"Username",
-			"Email",
-			"First Name",
-			"Last Name"
-		]
-		# Create list of form entries to get input later
-		# Then create and position label and entry widgets for form
-		self.formEntryList = []
-		for x in range(len(formFields)):
-			label = ctk.CTkLabel(formFieldsSection, text=formFields[x])
-			entry = ctk.CTkEntry(formFieldsSection)
-			label.grid(row=x, column=0, pady=10, ipadx=10)
-			entry.grid(row=x, column=1, pady=10, ipadx=10)
-			self.formEntryList.append(entry)
-
-		# Clear form button and then a confirm changes button
-		formBtnsSection = ctk.CTkFrame(form, fg_color="transparent")
-		clearFormBtn = ctk.CTkButton(formBtnsSection, text="Clear", command=lambda: clearEntryWidgets(self.formEntryList))
-		confirmEditsBtn = ctk.CTkButton(formBtnsSection, text="Confirm Edits")
-		openChangePasswordBtn = ctk.CTkButton(formBtnsSection, text="Change Password", command=lambda: self.master.openPage("changePasswordPage"))
-		openDeleteAccountBtn = ctk.CTkButton(formBtnsSection, text="Account Deletion", command=lambda: self.master.openPage("deleteAccountPage")) #type: ignore
-
-		# Structure the remaining page elements accordingly
-		formHeader.grid(row=0, column=0, padx=70, pady=10)
-		formHeading.grid(row=0, column=0)
-		self.formErrorMessage.grid(row=1, column=0)
-		formFieldsSection.grid(row=1, column=0, pady=10, ipadx=10)
-		formBtnsSection.grid(row=2, column=0, pady=10)
-		confirmEditsBtn.grid(row=0, column=0, padx=10, pady=10)
-		clearFormBtn.grid(row=1, column=0, padx=10, pady=10)
-		openChangePasswordBtn.grid(row=2, column=0, padx=10, pady=10)
-		openDeleteAccountBtn.grid(row=3, column=0, padx=10, pady=10)
-
