--- pages/editAvatarPage.py
@@ -1,11 +1,0 @@
-'''
-- Idea for this page: Have a grid of avatars and images. Should
-have a folder of image or image links where we open the urls
-and get the images. With each image we should have a button
-with text "select". If user presses the button, then we 
-change the style of the button to be noticeably different with 
-text "Current". This would update the avatar in the database. 
-
-
-
-'''
